+++ packages/better_player/lib/src/core/extensions/player_data_source_extension.dart
@@ -85,6 +85,11 @@ extension PlayerDataSourceExtension on BetterPlayerController {
           },
         ),
       );
+      if (_engine != null) {
+        _engine!.value = _engine!.value.copyWith(
+          errorDescription: exception.toString(),
+        );
+      }
       return;
     }